Calling [rope->method] now takes into account the _method param
Because web browsers do not allow form submissions to use any HTTP method but
GET and POST, it is necessary to allow for an alternative way to express the
actual method that was desired. This can now be done by setting _method as
either a URL, query or post parameter.