Adding a make install and improving command-line utility
Can now run a "make install" to install the LassoApp and the binary into the
correct folders in LASSO9_HOME. If LASSO9_HOME does not exist, it uses
/var/laasso/home as the default.

When the command-line utility was compiled, if there was a parse error in a
startup file, it would quit without printing any error information. (Run as a
script, it would print the information.) So I wrapped it in a [protect] block
and [handle_error] print the errors to stdout.