New project now checks LassoApp installation
When creating a new project, first it checks to make sure that the LassoApp
(rope.lassoapp) has been properly installed into either "LassoApps" or into
"LassoApps-disabled" directories.

If it has been installed in the latter directory, then it will copy the LassoApp
into the "lasso_home/LassoApps/" directory for the project.