Adding the ability to specify a method to call on the value of a rope->param
The rope->param methods are wrappers for web_request->param methods which return
their values as bytes. For the most part, this isn't helpful for most web apps,
so rope->param will return strings by default, but will also allow you to
specify the return type by passing the tag of a method to call on the (or each)
value(s). This will allow people to cast the value as integers, for example.