Adding orNullWrap, stringOrNull, integerOrNull, dateOrNull
The last three are meant to be passed to the "-as=" option in a rope->param
method call. They are meant to return NULL if nothing was passed, or an object
of the specified type using the value passed.

The orNullWrap method is meant to be an implementation detail, but one that
should make custom typeOrNull methods easy to write.