Changing how webapp gets loaded for Linux compatibility
We can't just symbolically link webapp to the LassoStartup folder because Lasso
doesn't sort the files there before loading them. While this is fine when running
OS X as it spits the files out in lexicographical order, Linux does not.

The solution is to symbolically link the webapp folder to the rope_webapp in the
LassoStartup folder and to stick a file in LassoStartup that will load all the
files in the correct order.